+++ CHANGELOG.md
@@ -11,6 +11,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - On a typical system during daylight hours, the Envoy will execute the task within 15 minutes.
 - More info about Power Production task here: https://support.enphase.com/s/article/How-do-I-disable-and-enable-power-production
 
+## [6.7.0] - (27.02.2023)
+## Changes
+- added powerfull RESTFull server to use with own automations
+- cleanup
+- config.schema updated
+
 ## [6.6.0] - (26.02.2023)
 ## Changes
 - added for ensemble summary Rest Power